[#66] Refactor : 가독성 향상을 위한 함수명 변경
